Added the application log to CopyToClipboard in CrossMgrAlien and CrossMgrImpinj Improved connection robustness for CrossMgrImpinj. Changed HextFormatToInt to return a long. Added screen refresh to DNSManager. Added screen refresh to RiderDetail after changing status. Updated CrossMgrImpinj version level.

+++ CrossMgrImpinj/pyllrp/pyllrp.py
@@ -631,7 +631,7 @@ def WaitForMessage( MessageID, ResponseClass, sock ):
 #-----------------------------------------------------------------------------
 
 def HexFormatToInt( value ):
-	return int(''.join( [ "%02X" % ord(x) for x in value ] ))
+	return long(''.join( [ "%02X" % ord(x) for x in value ] ))
 
 def GetBasicAddRospecMessage( MessageID = None, ROSpecID = 123, inventoryParameterSpecID = 1234, antennas = None ):
 	#-----------------------------------------------------------------------------
